requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes
agate==1.9.1
    # via
    #   dbt-adapters
    #   dbt-common
    #   dbt-core
annotated-types==0.7.0
    # via pydantic
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
babel==2.17.0
    # via agate
certifi==2025.8.3
    # via requests
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via
    #   dbt-core
    #   dbt-semantic-interfaces
colorama==0.4.6
    # via
    #   click
    #   dbt-common
    #   loguru
daff==1.4.2
    # via dbt-core
dbt-adapters==1.16.6
    # via dbt-core
dbt-common==1.29.0
    # via
    #   dbt-adapters
    #   dbt-core
dbt-core==1.10.11
    # via data-forge
dbt-extractor==0.6.0
    # via dbt-core
dbt-protos==1.0.348
    # via
    #   dbt-adapters
    #   dbt-common
    #   dbt-core
dbt-semantic-interfaces==0.9.0
    # via dbt-core
deepdiff==8.6.1
    # via dbt-common
idna==3.10
    # via requests
importlib-metadata==8.7.0
    # via dbt-semantic-interfaces
isodate==0.6.1
    # via
    #   agate
    #   dbt-common
isort==6.0.1
    # via data-forge
jinja2==3.1.6
    # via
    #   dbt-common
    #   dbt-core
    #   dbt-semantic-interfaces
jsonschema==4.25.1
    # via
    #   dbt-common
    #   dbt-core
    #   dbt-semantic-interfaces
jsonschema-specifications==2025.4.1
    # via jsonschema
leather==0.4.0
    # via agate
loguru==0.7.3
    # via data-forge
markupsafe==3.0.2
    # via jinja2
mashumaro==3.14
    # via
    #   dbt-adapters
    #   dbt-common
    #   dbt-core
more-itertools==10.8.0
    # via dbt-semantic-interfaces
msgpack==1.1.1
    # via mashumaro
networkx==3.5
    # via dbt-core
orderly-set==5.5.0
    # via deepdiff
packaging==25.0
    # via dbt-core
parsedatetime==2.6
    # via agate
pathspec==0.12.1
    # via
    #   dbt-common
    #   dbt-core
protobuf==6.32.0
    # via
    #   dbt-adapters
    #   dbt-common
    #   dbt-core
    #   dbt-protos
psycopg2-binary==2.9.10
    # via data-forge
pydantic==2.11.7
    # via
    #   dbt-core
    #   dbt-semantic-interfaces
pydantic-core==2.33.2
    # via pydantic
python-dateutil==2.9.0.post0
    # via
    #   dbt-common
    #   dbt-semantic-interfaces
python-dotenv==1.1.1
    # via data-forge
python-slugify==8.0.4
    # via agate
pytimeparse==1.1.8
    # via agate
pytz==2025.2
    # via
    #   dbt-adapters
    #   dbt-core
pyyaml==6.0.2
    # via
    #   dbt-core
    #   dbt-semantic-interfaces
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   data-forge
    #   dbt-common
    #   dbt-core
    #   snowplow-tracker
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via
    #   isodate
    #   python-dateutil
snowplow-tracker==1.1.0
    # via dbt-core
sqlparse==0.5.3
    # via dbt-core
text-unidecode==1.3
    # via python-slugify
typing-extensions==4.15.0
    # via
    #   dbt-adapters
    #   dbt-common
    #   dbt-core
    #   dbt-semantic-interfaces
    #   mashumaro
    #   pydantic
    #   pydantic-core
    #   snowplow-tracker
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2 ; sys_platform == 'win32'
    # via agate
urllib3==2.5.0
    # via requests
win32-setctime==1.2.0 ; sys_platform == 'win32'
    # via loguru
zipp==3.23.0
    # via importlib-metadata
